A tree control. Fix scrolling.
Isolate the treeview into its own TreeControl file so everyone can use
it. Fix scrolling while we're at it. Context menus still need to be
implemented.
